+++ test/test_unittest.py
@@ -30,7 +30,7 @@ class TestSchema(TestCase):
     result.stderr.re_match_lines(["Ran 1 test in.*", "OK"])
 
 
-def test_unittest_failure(testdir):
+def test_unittest_failure(testdir, hypothesis_above_6_54):
     module = testdir.make_test(
         """
 from unittest import TestCase
@@ -45,4 +45,8 @@ class TestSchema(TestCase):
     result = testdir.run(sys.executable, "-m", "unittest", str(module))
     assert result.ret == 1
     result.stderr.re_match_lines([".* assert 0.*", "FAILED (failures=1)"])
-    result.stdout.re_match_lines(["Falsifying example: .*"])
+    if hypothesis_above_6_54:
+        output = result.stderr
+    else:
+        output = result.stdout
+    output.re_match_lines(["Falsifying example: .*"])